issue-993: extract dashboard contract/runtime
