+++ crates/tau-coding-agent/Cargo.toml
@@ -35,6 +35,7 @@ tau-release-channel = { path = "../tau-release-channel" }
 tau-ops = { path = "../tau-ops" }
 tau-memory = { path = "../tau-memory" }
 tau-github-issues = { path = "../tau-github-issues" }
+tau-dashboard = { path = "../tau-dashboard" }
 tau-agent-core = { path = "../tau-agent-core" }
 tau-ai = { path = "../tau-ai" }
 reqwest.workspace = true
