+++ crates/tau-dashboard/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "tau-dashboard"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+tokio = { workspace = true, features = ["time"] }
+
+tau-core = { path = "../tau-core" }
+tau-runtime = { path = "../tau-runtime" }
+
+[dev-dependencies]
+tempfile = "3.10"
+tokio = { version = "1.43", features = ["rt-multi-thread", "macros", "time"] }
