+++ crates/tau-dashboard/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod dashboard_contract;
+pub mod dashboard_runtime;
